requirements.txt
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
ddt==1.2.1
decorator==4.3.2
docopt==0.6.2
fire==0.1.3
humanize==0.5.1
idna==2.8
logzero==1.5.0
lxml==4.3.2
Pillow==5.4.1
progress==1.5
pure-python-adb==0.1.6.dev0
py==1.8.0
requests==2.21.0
retry==0.9.2
six==1.12.0
tornado==6.0.1
uiautomator2==0.1.11
urllib3==1.24.1
weditor==0.1.1
whichcraft==0.5.2
